+++ deno.json
@@ -27,6 +27,6 @@
 	"lock": false,
 	"nodeModulesDir": true,
 	"test": {
-		"include": ["tests/*/*", "source/compiler/**", "source/wasm/**"]
+		"include": ["tests/**"]
 	}
 }
