+++ tests/e2e/compiler/fibonacci.test.ts
