+++ tests/e2e/compiler/numeric.test.ts
