+++ tests/e2e/wasm/hello-world.test.ts
@@ -1,6 +1,7 @@
 /// <reference lib="deno.ns" />
 import { fail, assertEquals } from "https://deno.land/std@0.201.0/assert/mod.ts";
-import { Module, Instruction, Type } from "../../source/wasm/index.ts";
+
+import { Module, Instruction, Type } from "~/wasm/index.ts";
 
 const decoder = new TextDecoder();
